Added player image, fixed enemy movement vectors and broke player movement

+++ src/BasicZombie.cpp
@@ -8,7 +8,7 @@
 
 #include "BasicZombie.hpp"
 
-BasicZombie::BasicZombie(int x, int y, int movementSpeed, int health, bool visible, std::string imgPath) : Creature(x, y, movementSpeed, health, visible, imgPath){
+BasicZombie::BasicZombie(int x, int y, float movementSpeed, int health, bool visible, std::string imgPath) : Creature(x, y, movementSpeed, health, visible, imgPath){
 }
