+++ src/Creature.hpp
@@ -15,8 +15,8 @@
 class Creature : public Entity{
 public:
     int health;
-    int movementSpeed;
-    Creature(int x, int y, int movementSpeed, int health, bool visible, std::string imgPath);
+    float movementSpeed;
+    Creature(int x, int y, float movementSpeed, int health, bool visible, std::string imgPath);
     int getHealth();
     void takeDamage(int dmg);
     void moveTo(float x, float y);
